separate predict_mode and implement inter-prediction (unfinished and not yet tested)

+++ src/bool_decoder.cc
@@ -82,7 +82,7 @@ int16_t BoolDecoder::Tree(const std::vector<uint8_t> &prob,
 
 uint32_t BoolDecoder::ReadUncoded(size_t n) {
   uint32_t res = 0;
-  for (size_t i = 0; i < n; ++i) res |= ReadByte() << (i << 3);
+  for (size_t i = 0; i < n; ++i) res |= uint32_t(ReadByte()) << (i << 3);
   return res;
 }
